Move plugin index.ts file
